Move service name info out of internal
